backend/requirements.txt
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
typer>=0.13.0
python-dotenv>=1.0.1
openai>=1.52.0
langchain>=0.2.12
langchain-openai>=0.1.21
langgraph>=0.1.19
langchain-core>=0.2.0
rich>=13.0.0
